moved getIndices to circuit-lib.js
