+++ circuit-lib/circuit-lib.js/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./merkle-tree";
+export * from "./indices";
