+++ zk.js/src/errors.ts
